- David H. Bronke merged changes from revisions 24643 through 24739 of 	  trunk into the streamloader branch.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/branches/soc/streamloader@24740 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/imesh/skeleton.h
@@ -283,7 +283,7 @@ struct iSkeletonFactory : public virtual iBase
 struct iSkeletonGraveyard : public virtual iBase
 {
   SCF_INTERFACE (iSkeletonGraveyard, 1, 0, 0);
-  virtual int GetFactoriesCount() = 0;
+  virtual size_t GetFactoriesCount() = 0;
   virtual iSkeletonFactory *CreateFactory(const char *name) = 0;
   virtual iSkeletonFactory *LoadFactory(const char *file_name) = 0;
   virtual iSkeletonFactory *FindFactory(const char *name) = 0;
